+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v8.3.0 (2024-09-02)
 
+- add support for snapshot queries (single use) (#215)
 - add support for `Query\Builder::whereNotInUnnest(...)` (#225)
 - `Query\Builder::whereIn` will now wrap values in `UNNEST` if the number of values exceeds the limit (950). (#)